Break out parseRedisUrl into its own file
